+++ examples/minifb_template.rs
@@ -17,13 +17,13 @@ const RENDER_HEIGHT: usize = 360;
 pub struct TemplateEngine {
     pub rasterizer: Rasterizer,
 
-    pub realtime: f32,
-    pub timescale: f32,
+    pub realtime: f64,
+    pub timescale: f64,
     pub tics: u64,
     pub fps: u64,
     pub fps_print: u64,
-    pub dt: f32,
-    pub dt_unscaled: f32,
+    pub dt: f64,
+    pub dt_unscaled: f64,
 
     dt_before: Instant,
 
@@ -81,7 +81,7 @@ impl TemplateEngine {
         // Image example for transparency
         let graphics_and_shit = Image::new("core/default.png");
 
-        let mut printtime: f32 = 0.0;
+        let mut printtime: f64 = 0.0;
 
         while window.is_open() && !window.is_key_down(Key::Escape) {
             self.update_times();
@@ -108,8 +108,8 @@ impl TemplateEngine {
             // Image drawing but T R A N S P A R E N T
             self.rasterizer.set_draw_mode(DrawMode::Alpha);
             self.rasterizer.opacity = 128;
-            self.rasterizer.pimg(&graphics_and_shit, 256 + ((self.realtime.cos()) * 128.0) as i32, 160);
-            self.rasterizer.pimg(&graphics_and_shit, 256 + ((-self.realtime.cos()) * 128.0) as i32, 160);
+            self.rasterizer.pimg(&graphics_and_shit, 256 + ((self.realtime.cos()) * 128.0) as i64, 160);
+            self.rasterizer.pimg(&graphics_and_shit, 256 + ((-self.realtime.cos()) * 128.0) as i64, 160);
             self.rasterizer.opacity = 255;
             self.rasterizer.set_draw_mode(DrawMode::Opaque);
 
@@ -138,8 +138,8 @@ impl TemplateEngine {
 
     pub fn update_times(&mut self) {
         let now = Instant::now();
-        let now_s = (now.elapsed().as_secs() as f32) + (now.elapsed().subsec_nanos() as f32 * 1.0e-9);
-        let before_s = (self.dt_before.elapsed().as_secs() as f32) + (self.dt_before.elapsed().subsec_nanos() as f32 * 1.0e-9);
+        let now_s = (now.elapsed().as_secs() as f64) + (now.elapsed().subsec_nanos() as f64 * 1.0e-9);
+        let before_s = (self.dt_before.elapsed().as_secs() as f64) + (self.dt_before.elapsed().subsec_nanos() as f64 * 1.0e-9);
         self.dt_unscaled = before_s - now_s;
         
         self.dt_before = Instant::now();
