+++ src/matrix4.rs
@@ -4,7 +4,7 @@ use crate::three::*;
 // Matrix 4x4 for 3D transformations
 #[derive(Copy, Clone, Debug)]
 pub struct Matrix4 {
-	pub m: [[f32; 4]; 4],
+	pub m: [[f64; 4]; 4],
 }
 
 impl Matrix4 {
@@ -37,11 +37,11 @@ impl Matrix4 {
 		nmtx
 	}
 
-	pub fn rotated(yaw: f32, pitch: f32, roll: f32) -> Matrix4 {
+	pub fn rotated(yaw: f64, pitch: f64, roll: f64) -> Matrix4 {
 		Matrix4::rotated_yaw(yaw) * Matrix4::rotated_pitch(pitch) *  Matrix4::rotated_roll(roll)
 	}
 
-	pub fn rotated_roll(theta: f32) -> Matrix4 {
+	pub fn rotated_roll(theta: f64) -> Matrix4 {
 		let sin = theta.sin();
 		let cos = theta.cos();
 
@@ -53,7 +53,7 @@ impl Matrix4 {
 		nmtx
 	}
 
-	pub fn rotated_pitch(theta: f32) -> Matrix4 {
+	pub fn rotated_pitch(theta: f64) -> Matrix4 {
 		let sin = theta.sin();
 		let cos = theta.cos();
 
@@ -65,7 +65,7 @@ impl Matrix4 {
 		nmtx
 	}
 
-	pub fn rotated_yaw(theta: f32) -> Matrix4 {
+	pub fn rotated_yaw(theta: f64) -> Matrix4 {
 		let sin = theta.sin();
 		let cos = theta.cos();
 
@@ -92,7 +92,7 @@ impl Matrix4 {
 		vo.y = vi.x * self.m[0][1] + vi.y * self.m[1][1] + vi.z * self.m[2][1] + self.m[3][1];
 		vo.z = vi.x * self.m[0][2] + vi.y * self.m[1][2] + vi.z * self.m[2][2] + self.m[3][2];
 
-		let w: f32 = vi.x * self.m[0][3] + vi.y * self.m[1][3] + vi.z * self.m[2][3] + self.m[3][3];
+		let w: f64 = vi.x * self.m[0][3] + vi.y * self.m[1][3] + vi.z * self.m[2][3] + self.m[3][3];
 
 		if w != 0.0 {
 			vo.x /= w; vo.y /= w; vo.z /= w;
